+++ odoo/local-src/sf_rma_sale_order/__manifest__.py
@@ -12,6 +12,7 @@
     "depends": [
         "sf_rma",
         "sf_rma_mrp_repair",
+        "sf_sale",
     ],
     "data": [
         "data/product.xml",