+++ HISTORY.rst
@@ -25,6 +25,7 @@ latest (unreleased)
 * BSSFL-531: Company logo
 * BSSFL-547: Add the module MRP Average cost
 * BSSFL-546: Repair time field
+* BSSFL-544: Repaired product is not invoiced
 
 **Bugfixes**